backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart==0.0.6
email-validator==2.3.0
pytest>=8.0.0
httpx>=0.23.0,<0.28.0
